Improve error log after incorrect SPF version.
When the TXT record is not a valid SPF record goSPF returns error.
However, for sake of better debugging, called should be informed which
record was returned.

This commit also adds a commit on why DNS related errors are returned to
the called and not handled in the
dns.LookupSPF() function.