Implement macro parsing.
Section 7 of RFC 7208 defines grammar for parsing macros used in >>exists<< and
>>exp<< terms. This patch is a squash of multiple commits from the >>macro<<
branch and intends to add big chunk of functionality merged with the master
branch.
I have decided to add separate  module (macro.go) and not mix Lexer/Parser.
Macro parser is almost identical with Lexer type, so is multiple methods, but
this is a tradeoff for not mixing two distinct lexers.