Use miekg/dns library.
Refactor whole code and use miekg's DNS library instead native go functions.
New library allows for more strict error code checking (to be done) as well as
easier testing (to be done as well).

This commit includes multiple changes:

*) Use new DNS library for querying DNS servers instead of net.Lookup*
functions.
*) Change how the errors are handled, now we have access to RCODE value
*) Move dns/dns.go to spf/dns.go
*) Remove dns directory
*) Add const variable spf.Nameserver pointing to the DNS nameserver we want to
query (this will be more configurable in the future)
*) Add NormalizeHost() function in dns.go that, after some sanity checks,
appends
a '.' character to the host name, as this is the format required by DNS
library.
*) Function checkHost does not rely on LookupSPF() function, the DNS call is
execute in the function body.
*) LookupSPF() and relevant tests were deleted.