Don't return Permerror upon RCODE return code.
We should pretend there were no matching results when the DNS call returns
NXDOMAIN (RCODE 3). This patch modifies Resolver.exchange() function to fix the
bug.

On the other hand, Permerror should be returned for the very first TXT lookup
of the <domain> in the check_host() function. As a result, interface Resolver
was extended with a method LookupTXTStrict, that looks for RCODE 3 DNS call
result and will return ErrDNSPermerror.

Close #26